swagger y readme actualizados

+++ src/controllers/users.controllers.js
@@ -13,10 +13,10 @@ export const register=[
   .trim()
   .notEmpty().withMessage("Completar este campo")
   .isEmail().withMessage("Formato incorrecto"),
-  body("admin")
+  body("role")
   .trim()
   .notEmpty().withMessage("Completar este campo")
-  .isBoolean().withMessage("El valor debe ser true o false"),
+  .matches(/^(admin|user)$/).withMessage("El rol debe ser admin o user"),
   body("confirmPassword")
   .trim()
   .custom((value,{req})=>{
@@ -32,8 +32,8 @@ export const register=[
     }
   try{
     const{
-      username,password,confirmPassword,email,admin}=req.body;
-    const user=await registerUser({username,password,confirmPassword,email,admin});
+      username,password,confirmPassword,email,role}=req.body;
+    const user=await registerUser({username,password,email,role});
     res.status(201).json(user)
   }catch(error){
     next(error)
@@ -50,9 +50,10 @@ export const login=async(req,res,next)=>{
 }
 export const getUsersController=async(req,res,next)=>{
   try{
-    if(req.user.admin){
+    if(req.user.role === "admin"){
       const users=await getUsers();
-      res.status(201).json(users);
+      res.status(201).json(users
+      );
     }else{
       throw new Error("El usuario no es un administrador")
     }
@@ -63,7 +64,7 @@ export const getUsersController=async(req,res,next)=>{
 export const getUserController=async (req,res,next)=>{
   try{
     const {id}=req.params;
-    if(req.user.admin){
+    if(req.user.role === "admin"){
       const user=await getUser(id)
       res.status(201).json(user);
     }else{
@@ -73,21 +74,43 @@ export const getUserController=async (req,res,next)=>{
     next(error)
   }
 }
-export const update=async(req,res,next)=>{
+export const update=[
+  body("username")
+  .trim()
+  .isLength({min:8,max:20}).withMessage("El usuario debe contener entre 8 y 20 caracteres"),
+  body("email")
+  .trim()
+  .isEmail().withMessage("Formato incorrecto"),
+  body("passworord")
+  .trim()
+  .isLength({min:8,max:20}).withMessage("Debe contener entre 8 y 20 caractéres"),
+  async(req,res,next)=>{
+    const errors=validationResult(req);
+    if(!errors.isEmpty()){
+      res.status(400).json({errors:errors.array()})
+    }
   try{
     const {id}=req.params
     const cambios=req.body
+    if(req.user.role === "admin"){
     const update=await updateUser(id,cambios)
     res.json(update)
+    }else{
+      throw new Ertor("El usuario no es un administrador")
+    }
   }catch(error){
     next(error)
   }
-}
+}]
 export const removeUser=async(req,res,next)=>{
   try{
     const {id}=req.params;
+    if(req.user.role === "admin"){
     const user=await deleteUser(id);
     res.json(user);
+    }else{
+     return res.status(401).json({error:"El usuario no es un administrador"});
+    }
   }catch(error){
     next(error)
   }
